refactor(runtime): move the endpoint spawn seam into runner/spawner.rs
runner.rs was 1501 production lines against a 1500 budget. main already sits
at 1498, so this is not really this branch's three added lines — the file is
at the edge and the next change to touch it trips the same gate.

Moved the spawn seam — NetworkEndpointSpawnRequest, the NetworkEndpointSpawner
trait, and RealNetworkEndpointSpawner — into the existing runner/ module tree
beside backend, refusal, sockets, and warm_claim. It is one cohesive unit: the
request, the trait it is passed to, and the single production impl. Re-exported
from runner so every caller's path is unchanged.

runner.rs is now 1459 lines, which leaves real headroom rather than one line.
check-uniform-vsock-egress still resolves the runner binding, so the claim-10
seam is unmoved.
